move the courseNameAndNumber folder to modals
